refactor forgot password page Relates #37
