+++ src/interfaces/User.ts
@@ -6,6 +6,7 @@ export interface User {
   lastLogin: string;
   userIp: string;
   userRoleId: number;
+  confirmationCode?: string;
 }
 export interface UserContext {
   user: User;
