+++ src/interfaces/forgotPasswordChildren.ts
@@ -1,0 +1,3 @@
+export default interface ForgotPasswordChildren {
+  handleComplete: () => void;
+}
